add rendering logic and state for form builder editing mode
