+++ src/Components/FormElements.js
@@ -130,7 +130,9 @@ export function RadioGroupField(props) {
     addLSMShemaRadioOption,
   });
 
-  const { name, label, placeholder, options, fielddata, ...rest } = props;
+  const { name, label, placeholder, options, fielddata, mode, ...rest } = props;
+
+  const isEditingMode = mode;
 
   let fieldLabel = label;
   if (!label) {
@@ -196,16 +198,20 @@ export function RadioGroupField(props) {
         </Field>
       </motion.div>
 
-      <EuiHorizontalRule />
+      {isEditingMode && (
+        <>
+          <EuiHorizontalRule />
 
-      <EuiButton
-        onClick={handleAddRadioOptionToField}
-        size="s"
-        color="text"
-        iconType={AddCircle}
-      >
-        Add option
-      </EuiButton>
+          <EuiButton
+            onClick={handleAddRadioOptionToField}
+            size="s"
+            color="text"
+            iconType={AddCircle}
+          >
+            Add option
+          </EuiButton>
+        </>
+      )}
     </>
   );
 }
@@ -216,7 +222,9 @@ export function CheckboxGroupField(props) {
     addLSMShemaCheckboxOption,
   });
 
-  const { name, label, placeholder, options, fielddata, ...rest } = props;
+  const { name, label, placeholder, options, fielddata, mode, ...rest } = props;
+
+  const isEditingMode = mode;
 
   let fieldLabel = label;
   if (!label) {
@@ -282,16 +290,20 @@ export function CheckboxGroupField(props) {
         </Field>
       </motion.div>
 
-      <EuiHorizontalRule />
+      {isEditingMode && (
+        <>
+          <EuiHorizontalRule />
 
-      <EuiButton
-        onClick={handleAddCheckboxOptionToField}
-        size="s"
-        color="text"
-        iconType={AddCircle}
-      >
-        Add option
-      </EuiButton>
+          <EuiButton
+            onClick={handleAddCheckboxOptionToField}
+            size="s"
+            color="text"
+            iconType={AddCircle}
+          >
+            Add option
+          </EuiButton>
+        </>
+      )}
     </>
   );
 }
